LayoutTests/fast/events/spatial-navigation/snav-multiple-select-expected.txt

2	
4		6
8	
FAIL gFocusedDocument.activeElement.getAttribute("id") should be 8. Was start.
PASS gFocusedDocument.activeElement.getAttribute("id") is "start"
FAIL gFocusedDocument.activeElement.getAttribute("id") should be 2. Was start.
PASS gFocusedDocument.activeElement.getAttribute("id") is "start"
PASS gFocusedDocument.activeElement.getAttribute("id") is "6"
PASS gFocusedDocument.activeElement.getAttribute("id") is "start"
PASS gFocusedDocument.activeElement.getAttribute("id") is "4"
PASS gFocusedDocument.activeElement.getAttribute("id") is "start"
